+++ src/Plainion.GraphViz.Tests/Properties/AssemblyInfo.cs
@@ -5,15 +5,15 @@ using System.Reflection;
 [assembly: AssemblyProductAttribute("Plainion.GraphViz")]
 [assembly: AssemblyDescriptionAttribute("Plainion.GraphViz")]
 [assembly: AssemblyCopyrightAttribute("Copyright @ 2019")]
-[assembly: AssemblyVersionAttribute("2.12.0")]
-[assembly: AssemblyFileVersionAttribute("2.12.0")]
+[assembly: AssemblyVersionAttribute("2.13.0")]
+[assembly: AssemblyFileVersionAttribute("2.13.0")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyTitle = "Plainion.GraphViz.Tests";
         internal const System.String AssemblyProduct = "Plainion.GraphViz";
         internal const System.String AssemblyDescription = "Plainion.GraphViz";
         internal const System.String AssemblyCopyright = "Copyright @ 2019";
-        internal const System.String AssemblyVersion = "2.12.0";
-        internal const System.String AssemblyFileVersion = "2.12.0";
+        internal const System.String AssemblyVersion = "2.13.0";
+        internal const System.String AssemblyFileVersion = "2.13.0";
     }
 }