+++ src/Plainion.GraphViz/Model/EdgeComparer.cs
@@ -1,0 +1,17 @@
+﻿using System.Collections.Generic;
+
+namespace Plainion.GraphViz.Model
+{
+    internal class EdgeComparer : IEqualityComparer<Edge>
+    {
+        public bool Equals(Edge x, Edge y)
+        {
+            return x.Id == y.Id;
+        }
+
+        public int GetHashCode(Edge obj)
+        {
+            return obj.Id.GetHashCode();
+        }
+    }
+}